refactor: move rollout bucketing into v4
The hash is shared by the v3 and v4 evaluation paths, so it lives in the
one that survives v3's removal rather than the one scheduled for deletion.

Its cross-library vectors move to src/testing alongside the specification
test server: support code that tests need and the package does not ship.
